Rename un-/deploy to remove/add
