+++ internal/cmd/alpha/remove/managed/managed.go
@@ -1,6 +1,8 @@
 package managed
 
 import (
+	"slices"
+
 	"github.com/kyma-project/cli.v3/internal/clierror"
 	"github.com/kyma-project/cli.v3/internal/cmdcommon"
 	"github.com/kyma-project/cli.v3/internal/kyma"
@@ -12,8 +14,7 @@ type managedConfig struct {
 	*cmdcommon.KymaConfig
 	cmdcommon.KubeClientConfig
 
-	module  string
-	channel string
+	module string
 }
 
 func NewManagedCMD(kymaConfig *cmdcommon.KymaConfig) *cobra.Command {
@@ -24,62 +25,47 @@ func NewManagedCMD(kymaConfig *cmdcommon.KymaConfig) *cobra.Command {
 
 	cmd := &cobra.Command{
 		Use:   "managed",
-		Short: "Deploy Kyma module on a managed Kyma instance",
+		Short: "Remove Kyma module on a managed Kyma instance",
 		PreRun: func(_ *cobra.Command, _ []string) {
 			clierror.Check(config.KubeClientConfig.Complete())
 		},
 		RunE: func(_ *cobra.Command, _ []string) error {
-			return runManagedDeploy(config)
+			return runRemoveManaged(config)
 		},
 	}
 
 	config.KubeClientConfig.AddFlag(cmd)
-	cmd.Flags().StringVar(&config.module, "module", "", "Name of the module to deploy")
-	cmd.Flags().StringVar(&config.channel, "channel", "", "Name of the Kyma channel to use for the module")
+	cmd.Flags().StringVar(&config.module, "module", "", "Name of the module to remove")
 
 	_ = cmd.MarkFlagRequired("module")
 
 	return cmd
 }
 
-func runManagedDeploy(config *managedConfig) error {
+func runRemoveManaged(config *managedConfig) error {
 	kymaCR, err := kyma.GetDefaultKyma(config.Ctx, config.KubeClient)
 	if err != nil {
 		return err
 	}
 
-	kymaCR = updateCR(kymaCR, config.module, config.channel)
+	kymaCR = updateCR(kymaCR, config.module)
 
 	_, err = kyma.UpdateDefaultKyma(config.Ctx, config.KubeClient, kymaCR)
 	return err
 }
 
-func updateCR(kymaCR *unstructured.Unstructured, moduleName, moduleChannel string) *unstructured.Unstructured {
+func updateCR(kymaCR *unstructured.Unstructured, moduleName string) *unstructured.Unstructured {
 	newCR := kymaCR.DeepCopy()
 	spec := newCR.Object["spec"].(map[string]interface{})
-	modules := make([]kyma.Module, 0)
+	modules := []kyma.Module{}
 	for _, m := range spec["modules"].([]interface{}) {
 		modules = append(modules, kyma.ModuleFromInterface(m.(map[string]interface{})))
 	}
 
-	moduleExists := false
-	for i, m := range modules {
-		if m.Name == moduleName {
-			// module already exists, update channel
-			modules[i].Channel = moduleChannel
-			moduleExists = true
-			break
-		}
-	}
-
-	if !moduleExists {
-		modules = append(modules, kyma.Module{
-			Name:    moduleName,
-			Channel: moduleChannel,
-		})
-	}
+	modules = slices.DeleteFunc(modules, func(m kyma.Module) bool {
+		return m.Name == moduleName
+	})
 
 	spec["modules"] = modules
-
 	return newCR
 }
